Rewrite the API to use protocols
- Closes #69
- Closes #61
- Closes #22
