+++ Source/AnimatorDelegate.swift
@@ -1,0 +1,4 @@
+/// The protocol that an animator delegate needs to conform to.
+public protocol AnimatorDelegate: class {
+  func animatorHasNewFrame()
+}
