+++ Source/Info.plist
@@ -19,7 +19,7 @@
 	<key>CFBundleSignature</key>
 	<string>????</string>
 	<key>CFBundleVersion</key>
-	<string>102</string>
+	<string>118</string>
 	<key>NSPrincipalClass</key>
 	<string></string>
 </dict>
